refactor: split admin tenant client APIs
